+++ CHANGELOG.md
@@ -7,6 +7,13 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Changed SMTP_USERNAME and SMTP_PASSWORD to SENSU_EMAIL_SMTP_USERNAME and SENSU_EMAIL_SMTP_PASSWORD for consistency
+- Changed some error returns to be more consistent
+
+### Added
+- Added support for annotations for certain variables
+
 ## [0.1.0] - 2019-03-04
 
 ### Changed