+++ CHANGELOG.md
@@ -7,6 +7,15 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Updated documentation to reflect loginauth and annotations
+
+### Added
+- Added loginauth auth mechanism
+- Added Content-Type to support HTML emails
+- Added plugin Keyspace so as to support configuraion overrides with annotations
+- Added subject templating via argument/annotation
+
 ## [0.1.0] - 2019-03-04
 
 ### Changed